bridge/requirements.txt
bottle
pyserial